Extract the SocketContainer
